# docs/contracts/openapi-v1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Merge Game API",
    "description": "Unity \uBA38\uC9C0 \uAC8C\uC784 \uD074\uB77C\uC774\uC5B8\uD2B8\uB97C \uC704\uD55C \uC11C\uBC84 \uAD8C\uC704\uD615 HTTP API\uC785\uB2C8\uB2E4.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/board": {
      "post": {
        "tags": [
          "Board"
        ],
        "summary": "\uD50C\uB808\uC774\uC5B4 \uBCF4\uB4DC\uB97C \uCD5C\uCD08 \uD55C \uBC88 \uC0DD\uC131\uD558\uACE0 \uC774\uBBF8 \uC788\uC73C\uBA74 \uD604\uC7AC \uC0C1\uD0DC\uB97C \uADF8\uB300\uB85C \uBC18\uD658\uD569\uB2C8\uB2E4.",
        "operationId": "InitializeBoard",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardState"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardState"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardErrorResponse"
                }
              }
            }
          }
        },
        "deprecated": true,
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "get": {
        "tags": [
          "Board"
        ],
        "summary": "\uD604\uC7AC revision\uACFC \uC2AC\uB86F\uBCC4 \uC544\uC774\uD15C\uC744 \uD3EC\uD568\uD55C \uC804\uCCB4 \uBCF4\uB4DC\uB97C \uBC18\uD658\uD569\uB2C8\uB2E4.",
        "operationId": "GetBoard",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardState"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/board/merge": {
      "post": {
        "tags": [
          "Board"
        ],
        "summary": "\uB450 \uC2AC\uB86F\uC744 \uC11C\uBC84\uC5D0\uC11C \uAC80\uC99D\uD558\uACE0 \uC131\uACF5\uD558\uBA74 source\uB97C \uC18C\uBE44\uD574 target\uC744 \uB2E4\uC74C \uB2E8\uACC4\uB85C \uC62C\uB9BD\uB2C8\uB2E4.",
        "operationId": "MergeBoardItems",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeBoardItemsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardState"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/board/actions": {
      "post": {
        "tags": [
          "Board"
        ],
        "summary": "\uB4DC\uB798\uADF8\uD55C \uB450 \uC2AC\uB86F\uC744 \uC11C\uBC84 \uC0C1\uD0DC\uC5D0 \uB530\uB77C \uC774\uB3D9, \uBA38\uC9C0 \uB610\uB294 \uAD50\uD658\uC73C\uB85C \uD310\uC815\uD569\uB2C8\uB2E4.",
        "operationId": "ApplyBoardAction",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyBoardActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/board/items/{itemId}/sell": {
      "post": {
        "tags": [
          "Board"
        ],
        "summary": "\uC11C\uBC84 \uCE74\uD0C8\uB85C\uADF8 \uAC00\uACA9\uC73C\uB85C \uC544\uC774\uD15C\uC744 \uC81C\uAC70\uD558\uACE0 \uCF54\uC778\uC744 \uC6D0\uC790\uC801\uC73C\uB85C \uC9C0\uAE09\uD569\uB2C8\uB2E4.",
        "operationId": "SellBoardItem",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SellBoardItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SellBoardItemResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardItemSaleErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardItemSaleErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardItemSaleErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardItemSaleErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/board/generators/{generatorId}/produce": {
      "post": {
        "tags": [
          "Board"
        ],
        "summary": "\uC11C\uBC84\uAC00 \uC0DD\uC131 \uC544\uC774\uD15C\uACFC \uBE48 \uC2AC\uB86F\uC744 \uC120\uD0DD\uD558\uACE0 \uC5D0\uB108\uC9C0\u00B7\uBCF4\uB4DC\u00B7\uCDA9\uC804 \uC0C1\uD0DC\uB97C \uC6D0\uC790\uC801\uC73C\uB85C \uBCC0\uACBD\uD569\uB2C8\uB2E4.\r\n\uAC19\uC740 \uBA71\uB4F1 \uD0A4\uC758 \uC131\uACF5 \uC694\uCCAD\uC740 \uC800\uC7A5\uB41C \uCD5C\uCD08 \uC751\uB2F5\uB9CC \uC7AC\uC0DD\uD569\uB2C8\uB2E4.",
        "operationId": "ProduceGeneratorItem",
        "parameters": [
          {
            "name": "generatorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProduceGeneratorItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratorProduceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratorProduceErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratorProduceErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratorProduceErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratorProduceErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApplyBoardActionRequest": {
        "type": "object",
        "properties": {
          "sourceSlot": {
            "type": "integer",
            "format": "int32"
          },
          "targetSlot": {
            "type": "integer",
            "format": "int32"
          },
          "expectedBoardRevision": {
            "type": "integer",
            "format": "int64"
          },
          "idempotencyKey": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "\uD074\uB77C\uC774\uC5B8\uD2B8\uAC00 \uB4DC\uB798\uADF8\uD55C \uB450 \uC2AC\uB86F\uACFC \uB9C8\uC9C0\uB9C9\uC73C\uB85C \uD655\uC778\uD55C revision\uB9CC \uC804\uB2EC\uD569\uB2C8\uB2E4."
      },
...
      "BoardActionResponse": {
        "type": "object",
        "properties": {
          "board": {
            "$ref": "#/components/schemas/BoardState"
          },
          "action": {
            "type": "string",
            "nullable": true
          },
          "sourceSlot": {
            "type": "integer",
            "format": "int32"
          },
          "targetSlot": {
            "type": "integer",
            "format": "int32"
          },
          "resultItem": {
            "$ref": "#/components/schemas/BoardItemState"
          },
          "replayed": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "BoardErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "currentRevision": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "board": {
            "$ref": "#/components/schemas/BoardState"
          }
        },
        "additionalProperties": false,
        "description": "\uBCF4\uB4DC API \uC2E4\uD328 \uC2DC \uC548\uC815\uC801\uC778 \uCF54\uB4DC\uC640 \uCD5C\uC2E0 \uB3D9\uAE30\uD654 \uC815\uBCF4\uB97C \uBC18\uD658\uD569\uB2C8\uB2E4."
      },
...
      "BoardState": {
        "type": "object",
        "properties": {
          "playerId": {
            "type": "string",
            "description": "\uBCF4\uB4DC \uC18C\uC720 \uD50C\uB808\uC774\uC5B4 ID\uC785\uB2C8\uB2E4.",
            "format": "uuid"
          },
          "width": {
            "type": "integer",
            "description": "\uBCF4\uB4DC \uAC00\uB85C \uC2AC\uB86F \uC218\uC785\uB2C8\uB2E4.",
            "format": "int32"
          },
          "height": {
            "type": "integer",
            "description": "\uBCF4\uB4DC \uC138\uB85C \uC2AC\uB86F \uC218\uC785\uB2C8\uB2E4.",
            "format": "int32"
          },
          "revision": {
            "type": "integer",
            "description": "\uB2E4\uC74C \uBCC0\uACBD \uC694\uCCAD\uC5D0\uC11C expectedRevision\uC73C\uB85C \uBCF4\uB0BC \uBC84\uC804\uC785\uB2C8\uB2E4.",
            "format": "int64"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BoardItemState"
            },
            "description": "\uC2AC\uB86F \uC21C\uC11C\uB85C \uC815\uB82C\uB41C \uD604\uC7AC \uC544\uC774\uD15C \uBAA9\uB85D\uC785\uB2C8\uB2E4.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "HTTP \uACC4\uCE35\uC5D0 \uC804\uB2EC\uD560 \uD604\uC7AC \uBCF4\uB4DC \uC804\uCCB4 \uC0C1\uD0DC\uC785\uB2C8\uB2E4."
      },
...
      "MergeBoardItemsRequest": {
        "type": "object",
        "properties": {
          "sourceSlot": {
            "type": "integer",
            "description": "\uC131\uACF5 \uC2DC \uC18C\uBE44\uB418\uC5B4 \uBE44\uAC8C \uB418\uB294 \uC6D0\uBCF8 \uC2AC\uB86F\uC785\uB2C8\uB2E4.",
            "format": "int32"
          },
          "targetSlot": {
            "type": "integer",
            "description": "\uC131\uACF5 \uC2DC \uB2E4\uC74C \uB808\uBCA8 \uC544\uC774\uD15C\uC774 \uB0A8\uB294 \uB300\uC0C1 \uC2AC\uB86F\uC785\uB2C8\uB2E4.",
            "format": "int32"
          },
          "expectedRevision": {
            "type": "integer",
            "description": "\uD074\uB77C\uC774\uC5B8\uD2B8\uAC00 \uB9C8\uC9C0\uB9C9\uC73C\uB85C \uD655\uC778\uD55C \uBCF4\uB4DC revision\uC785\uB2C8\uB2E4.",
            "format": "int64"
          }
        },
        "additionalProperties": false,
        "description": "\uB450 \uC2AC\uB86F \uBA38\uC9C0 \uC694\uCCAD\uC785\uB2C8\uB2E4."
      },
...
      "SellBoardItemRequest": {
        "type": "object",
        "properties": {
          "expectedBoardRevision": {
            "type": "integer",
            "format": "int64"
          },
          "expectedEconomyRevision": {
            "type": "integer",
            "format": "int64"
          },
          "idempotencyKey": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "\uD310\uB9E4 \uC2DC \uD074\uB77C\uC774\uC5B8\uD2B8\uAC00 \uD655\uC778\uD55C \uBCF4\uB4DC\u00B7\uACBD\uC81C revision\uACFC \uBA71\uB4F1 \uD0A4\uB97C \uC804\uB2EC\uD569\uB2C8\uB2E4."
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "POST /api/v1/auth/guest\uC5D0\uC11C \uBC1C\uAE09\uBC1B\uC740 accessToken\uC744 \uC785\uB825\uD569\uB2C8\uB2E4.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "AdminApiKey": {
        "type": "apiKey",
        "description": "\uC6B4\uC601 \uBE44\uBC00 \uC800\uC7A5\uC18C\uC5D0\uC11C \uC8FC\uC785\uD55C \uAD00\uB9AC\uC790 API \uD0A4\uC785\uB2C8\uB2E4. Unity \uD074\uB77C\uC774\uC5B8\uD2B8\uC5D0\uB294 \uD3EC\uD568\uD558\uC9C0 \uC54A\uC2B5\uB2C8\uB2E4.",
        "name": "X-Admin-Key",
        "in": "header"
      }
    }
  }
}